Move AppEnvironment to separate file
