+++ Passepartout/Shared/AppEnvironment.swift
@@ -1,0 +1,53 @@
+//
+//  AppEnvironment.swift
+//  Passepartout
+//
+//  Created by Davide De Rosa on 11/27/24.
+//  Copyright (c) 2024 Davide De Rosa. All rights reserved.
+//
+//  https://github.com/passepartoutvpn
+//
+//  This file is part of Passepartout.
+//
+//  Passepartout is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  Passepartout is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
+//
+
+import CommonIAP
+import Foundation
+import PassepartoutKit
+
+enum AppEnvironment {
+    static var isFakeIAP: Bool {
+        ProcessInfo.processInfo.environment["PP_FAKE_IAP"] == "1"
+    }
+
+    static var isFakeMigration: Bool {
+        ProcessInfo.processInfo.environment["PP_FAKE_MIGRATION"] == "1"
+    }
+
+    static var userLevel: AppUserLevel? {
+        if let envString = ProcessInfo.processInfo.environment["PP_USER_LEVEL"],
+           let envValue = Int(envString),
+           let testAppType = AppUserLevel(rawValue: envValue) {
+
+            return testAppType
+        }
+        if let infoValue = BundleConfiguration.mainIntegerIfPresent(for: .userLevel),
+           let testAppType = AppUserLevel(rawValue: infoValue) {
+
+            return testAppType
+        }
+        return nil
+    }
+}
